перенес useState в компонент offer-card
