+++ src/components/offers-list/offers-list.jsx
@@ -1,19 +1,10 @@
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import OfferCard from '../offer-card/offer-card';
 import offerTypes from '../../prop-types/offer-types';
 
 const OffersList = (props) => {
   const {offers} = props;
-  const [activeOffer, setActiveOffer] = useState(null);
-
-  const onHover = (id) => {
-    setActiveOffer(id);
-  };
-
-  useEffect(() => {
-    document.title = `Предложение ${activeOffer}`;
-  });
 
   return (
     <>
@@ -21,7 +12,6 @@ const OffersList = (props) => {
         <OfferCard
           key={id}
           offer={offer}
-          onHover={onHover}
         />
       )}
     </>
